move pipelines/util to util/traceback_util
