+++ depobs/scanner/docker/containers.py
@@ -28,7 +28,7 @@ import aiodocker
 from depobs.scanner.docker.client import aiodocker_client
 import depobs.scanner.docker.log_reader as docker_log_reader
 from depobs.scanner.models.git_ref import GitRef, GitRefKind
-from depobs.scanner.pipelines.util import exc_to_str
+from depobs.util.traceback_util import exc_to_str
 
 log = logging.getLogger(__name__)
