+++ depobs/scanner/docker/images.py
@@ -4,7 +4,7 @@ from typing import Iterable
 
 import depobs.scanner.docker.containers as containers
 from depobs.scanner.models.docker_image import DockerImage
-from depobs.scanner.pipelines.util import exc_to_str
+from depobs.util.traceback_util import exc_to_str
 
 log = logging.getLogger(__name__)
